ex1_first_test/requirements.txt
langchain-sqlserver
python-dotenv
pypdf
langchain-openai
langchain
langchain-core
langchain-text-splitters
langchain-community
langgraph
